Reorganize slightly
